requirements.txt
africastalking==1.2.8
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
django-filter==24.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
idna==3.7
phonenumbers==8.13.40
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.32.3
schema==0.7.7
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
